+++ tests/test_datasets/test_transforms/test_converting.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from copy import deepcopy
+
 from unittest import TestCase
 
 import numpy as np
@@ -107,30 +107,3 @@ class TestKeypointConverter(TestCase):
             self.assertTrue(
                 (results['keypoints_visible'][:, target_index, 0] ==
                  self.data_info['keypoints_visible'][:, source_index]).all())
-
-    def test_transform_sigmas(self):
-
-        mapping = [(3, 0), (6, 1), (16, 2), (5, 3)]
-        transform = KeypointConverter(num_keypoints=5, mapping=mapping)
-        sigmas = np.random.rand(17)
-        new_sigmas = transform.transform_sigmas(sigmas)
-        self.assertEqual(len(new_sigmas), 5)
-        for i, j in mapping:
-            self.assertEqual(sigmas[i], new_sigmas[j])
-
-    def test_transform_ann(self):
-        mapping = [(3, 0), (6, 1), (16, 2), (5, 3)]
-        transform = KeypointConverter(num_keypoints=5, mapping=mapping)
-
-        ann_info = dict(
-            num_keypoints=17,
-            keypoints=np.random.randint(3, size=(17 * 3, )).tolist())
-        ann_info_copy = deepcopy(ann_info)
-
-        _ = transform.transform_ann(ann_info)
-
-        self.assertEqual(ann_info['num_keypoints'], 5)
-        self.assertEqual(len(ann_info['keypoints']), 15)
-        for i, j in mapping:
-            self.assertListEqual(ann_info_copy['keypoints'][i * 3:i * 3 + 3],
-                                 ann_info['keypoints'][j * 3:j * 3 + 3])
